requirements.txt
torch
torchvision
timm
streamlit
opencv-python
albumentations
scikit-learn
pillow
